Adds temporary directories for sample creation
