--- Streaming-old/ImplicitStreams/ImplicitStreams.Client/ImplicitStreams.Client.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Orleans.Client" Version="8.2.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\ImplicitStreams.Shared\ImplicitStreams.Shared.csproj" />
-    </ItemGroup>
-
-</Project>
