--- Streaming-old/ImplicitStreams/ImplicitStreams.Shared/Constants.cs
@@ -1,7 +1,0 @@
-namespace ImplicitStreams.Shared;
-
-public static class Constants
-{
-    public const string StreamProvider = "StreamProvider";
-    public const string StreamNamespace = "COUNTER";
-}
