--- Streaming-old/ImplicitStreams/ImplicitStreams.Shared/IConsumerGrain.cs
@@ -1,6 +1,0 @@
-namespace ImplicitStreams.Shared;
-
-public interface IConsumerGrain : IGrainWithGuidKey
-{
-    ValueTask<int> GetCount();
-}
