--- Streaming-old/ImplicitStreams/ImplicitStreams.Shared/IProducerGrain.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.VisualBasic;
-
-namespace ImplicitStreams.Shared;
-
-public interface IProducerGrain : IGrainWithGuidKey
-{
-    Task StartProducing();
-}
