--- Streaming-old/ImplicitStreams/ImplicitStreams.Shared/ImplicitStreams.Shared.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Orleans.Sdk" Version="8.2.0" />
-      <PackageReference Include="Microsoft.Orleans.Streaming" Version="8.2.0" />
-    </ItemGroup>
-
-</Project>
