--- Streaming-old/ImplicitStreams/ImplicitStreams.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImplicitStreams.SiloHost", "ImplicitStreams.SiloHost\ImplicitStreams.SiloHost.csproj", "{281906AD-D2AE-470E-B6AD-E063C2B14183}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImplicitStreams.Shared", "ImplicitStreams.Shared\ImplicitStreams.Shared.csproj", "{F2F6C1A1-D0AC-4395-B19A-7EB490C1899E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImplicitStreams.Client", "ImplicitStreams.Client\ImplicitStreams.Client.csproj", "{A474C445-4D30-45CE-8B27-94B1D02B3C55}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{281906AD-D2AE-470E-B6AD-E063C2B14183}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{281906AD-D2AE-470E-B6AD-E063C2B14183}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{281906AD-D2AE-470E-B6AD-E063C2B14183}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{281906AD-D2AE-470E-B6AD-E063C2B14183}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F2F6C1A1-D0AC-4395-B19A-7EB490C1899E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F2F6C1A1-D0AC-4395-B19A-7EB490C1899E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F2F6C1A1-D0AC-4395-B19A-7EB490C1899E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F2F6C1A1-D0AC-4395-B19A-7EB490C1899E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A474C445-4D30-45CE-8B27-94B1D02B3C55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A474C445-4D30-45CE-8B27-94B1D02B3C55}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A474C445-4D30-45CE-8B27-94B1D02B3C55}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A474C445-4D30-45CE-8B27-94B1D02B3C55}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
