+++ patterns/patterns-satellite/README.md
@@ -1,0 +1,1 @@
+# Satellite Pattern
