+++ streaming/streaming-implicit/README.md
@@ -1,0 +1,1 @@
+# Implicit Stream Consumers
