+++ streaming/streaming-storage-queues/README.md
@@ -1,0 +1,1 @@
+# Streaming with Azure Storage Queues
